moved Test into its own component
